Create `ShortURLConverter` class
Extract URL shortening functionality into a new class. This moves the
work outside of the `ShortURL` template tag class and into a new
`ShortURLConverter` class.

This allows us to ask for short URLs outside the context of the template
tag.

It now becomes possible to find the short URL of an object in other
places, like an API.